+++ crates/aiken/src/cmd/blueprint/policy.rs
@@ -37,19 +37,11 @@ pub fn exec(
     }: Args,
 ) -> miette::Result<()> {
     with_project(directory.as_deref(), false, false, |p| {
-        let title = module.as_ref().map(|m| {
-            format!(
-                "{m}{}",
-                validator
-                    .as_ref()
-                    .map(|v| format!(".{v}"))
-                    .unwrap_or_default()
-            )
-        });
-
-        let title = title.as_ref().or(validator.as_ref());
-
-        let policy = p.policy(title, p.blueprint_path(input.as_deref()).as_path())?;
+        let policy = p.policy(
+            module.as_deref(),
+            validator.as_deref(),
+            p.blueprint_path(input.as_deref()).as_path(),
+        )?;
 
         println!("{}", policy);